route-table: Use classifier versioning to replace all routes atomically.
The route table is implemented in terms of a classifier structure, which
supports versioning with atomic transactional updates, but OVS didn't
actually use it for the route table.  This commit starts using that feature
for the route table and makes OVS atomically replace one set of routes with
another.  This should fix a reported bug in which the routing table was
periodically empty due to refreshes.

Reported-by: Yinpeijun <yinpeijun@huawei.com>
Reported-at: https://mail.openvswitch.org/pipermail/ovs-discuss/2018-February/046144.html
Signed-off-by: Ben Pfaff <blp@ovn.org>