[RFC] ovn-controller: Experiment with restricting access to columns.
To make ovn-controller recompute incrementally, we need accurate
dependencies for each function that reads or writes a table.  It's
currently difficult to be sure about these dependencies, and certainly
difficult to maintain them over time, because there's no way to actually
enforce them.

This commit experiments with an approach that allows for fairly
fine-grained access control within ovn-controller to tables and columns.
It's based on generating a new version of the IDL data structures for each
case where we want different access control.  All of these data structures
have the same format, but the columns that a given piece of code is not
supposed to touch are renamed to discourage programmers from using them,
e.g. they're given names suffixed with "__noaccess" or "__writeonly".
(This means that there is no runtime overhead to the access control since
it only requires a cast to convert between the regular and restricted
versions.) In addition, when a columns is supposed to be read-only,
functions for modifying the column are not supplied.

Type-safe indexes are supplied in the same way.

Comments?

CC: Han Zhou <zhouhan@gmail.com>
Signed-off-by: Ben Pfaff <blp@ovn.org>