new ddlog string syntax
Placing a dollar sign in front of a string is ugly and extremely
error-prone (since the compiler does not complain if you forget it).
Therefore this extra `$` has been removed in the the new ddlog syntax,
which now treats all quoted strings as interpolated strings, so now
you can write

`"Value of x is ${x}"`

instead of

`$"Value of x is ${x}"`.

In addition, interpolated string literals are automatically concatenated
now:

`"x is ${x}, " "y is ${y}"`

is equivalent to

`"x is ${x}, y is ${y}"`