configure: Improve DDlog configure-time checks.
The configure-time check accepted any directory as the argument to
--with-ddlog, but only the "lib" directory in a differential-datalog
checkout would work.  This commit checks that std.dl exists in the
directory, to make it more likely that the user won't get an error later
in the build process.

The configure code here was printing "yes" or "no" on a line by itself.
This commit makes it printing "checking for DDlog library directory..."
just before that, and makes it print the configured directory or "none"
instead of "yes" or "no".

This commit adapts the configure code so that it adheres more closely
to the usual conventions, e.g. for variable naming.  It also improves
the --help text.

Signed-off-by: Ben Pfaff <blp@ovn.org>