northd: Split up automake.mk to improve ovsdb2ddlog dependencies.
With all of the configuration for generating the DDlog files for OVSDB
schemas in the makefile itself, there's not a graceful way to
regenerate the DDlog when the configuration changes (e.g. because some
new field or table needs to be handled) because it's too expensive to
regenerate them whenever the Makefile changes at all.  But by putting
the configuration in a separate script, we can add a more granular
dependency.

Signed-off-by: Ben Pfaff <blp@ovn.org>