configure: Use autotools idioms to check for ddlog-related programs.
Usually, when one checks for a program in an autotools context, one
also gives the user a chance to specify an alternate program to execute,
by setting an appropriate variable.  This commit makes that happen.

This also adds a check for the rust compiler and makes it possible for
the user to easily override it at configure time.  I seem to recall that
I'd somehow installed cargo but not rustc or something stupid like that,
and so this actually helped me.

Signed-off-by: Ben Pfaff <blp@ovn.org>