northd: Support new fields in the northbound and southbound schemas.
A ddlog rule must initialize every field in the head data structure.
This commit adds the fields that were missing in a few tables that had
been augmented without the .dl files receiving updates.

In the northbound database, we don't write to NB_Global.name, so mark
it read-only and don't otherwise worry about it.

Signed-off-by: Ben Pfaff <blp@ovn.org>