northd-ddlog: Avoid variable/function name clash.
An upcoming version of DDlog will introduce support for closures.  As
part of this change, variables and functions will share the same
namespace, with variable declarations overriding function names in their
local scope.  This causes some existing code fail to type check.
Specifically, this happens in functions and rules that use a function
named `f` while declaring a variable with the same name.  The
fix is to rename such variables.