atlocal: Drop support for glibc before version 2.11.
The "ldd" call here didn't work if libtool was involved and would print
an error message.  We could fix that, but the check is only needed for
glibc earlier than 2.11.  glibc 2.11 was released in 2009, so it should
be safe to expect that testers are running it or a newer version.

Signed-off-by: Ben Pfaff <blp@ovn.org>