ovn-northd-ddlog: Improve type safety for datapath stages.
Until now, whether a logical flow stage was for a logical router or
switch, whether it was ingress or egress, and what particular stage it
was were each separate attributes.  It was possible to create an
invalid stage, e.g. a "logical router" stage that was actually present
only in logical switches.  This would result in a bug.

This commit makes such bugs much harder to cause.  They will generally
become compile-time errors.

This code refactoring shouldn't change ovn-northd-ddlog behavior.

Signed-off-by: Ben Pfaff <blp@ovn.org>