ovn-northd-ddlog: Make map_get_*() more object-like.
These functions were introduced before the object-like invocation
form was added to DDlog.  It's usually easier to read the object-like
form, so this changes the functions to this form and updates all the
callers.

It seemed to me that map_get_int_def_limit() was more general if we
just introduced a general-purpose clamp() function, so I did that.

Signed-off-by: Ben Pfaff <blp@ovn.org>