ovn_northd.dl: Save memory with streaming group_by
This is the first step towards optimizing the memory usage of
`Flow` relation and relations derived from it.

Profiling showed that a lot of memory is used in the output part of the
pipeline where the `Flow` relation is transformed into `Out_Logical_Flow`
to be sent to OVSDB.  `Flow` is a very large relation that dwarfs all
other relations in the program.  Rules that join and aggregate `Flow`
and its derivatives require DDlog to create several arrangements of
these relations, which end up dominating the memory footprint of the
program.

In the following rule, `Flow` is first arranged in preparation for
grouping.  The `group_by` operator internally creates another
arrangement with a record for each group.  The final join with
`UseLogicalDatapathGroups` arranges the result of `group_by` again.

```
AggregatedFlow(.logical_datapaths = g.to_set(),
               .stage = stage,
               .priority = priority,
               .__match = __match,
               .actions = actions,
               .external_ids = external_ids) :-
    Flow(logical_datapath, stage, priority, __match, actions, external_ids),
    var g = logical_datapath.group_by((stage, priority, __match, actions, external_ids)),
    UseLogicalDatapathGroups[true].
```

The next rule creates another arrangement of `Flow`:

```
AggregatedFlow(.logical_datapaths = set_singleton(logical_datapath),
    Flow(logical_datapath, stage, priority, __match, actions, external_ids),
    UseLogicalDatapathGroups[false].
```

`AggregatedFlow` is used to compute `Out_Logical_Flow`.  Since the
latter is declared as `output relation`, DDlog applies the `distinct`
operator to it to eliminate duplicate insertions and deletions, which
creates an arrangement internally.

Finally, the index over `Out_Logical_Flow` is backed by yet another
relation.

In this commit we start alleviating these overheads.  We temporarily
remove the `UseLogicalDatapathGroups` switch to simplify things, so
that datapath grouping is always enabled.  We'll bring
`UseLogicalDatapathGroups` back in a future commit.  The most important
change is the use of a new aggregation technique that avoids having to
arrange the entire relation before grouping it.  This relies on the use
of two recently added DDlog operators: delay and differentiation.  These
operators and the aggregation method using them are documented here:
https://github.com/vmware/differential-datalog/blob/master/doc/tutorial/tutorial.md#streaming-aggregation-differentiation-and-delay-operators
See comments in the source code for more details.

Signed-off-by: Leonid Ryzhyk <lryzhyk@vmware.com>