ovn_northd.dl: Make Out_Logical_Flows a multiset.
We eliminate another redundant arrangement.  DDlog uses differential
dataflow's `distinct` operator (more precisely, it uses
`threshold_total`, but the semantics is the same) to eliminate
duplicates in output relations.  This guarantees that the client won't
see e.g., multiple insertions followed by a deletion.  Handling such
outputs would require tracking the weights of output records in
`ovn-northd-ddlog.c`.  Internally `threshold_total` creates an
arrangement of the relation.  `threshold_total` is not necessary if the
relation is distinct (does not contain duplicates) by construction.
This is the case for `Out_Logical_Flows`, but the compiler is not clever
enough to figure this out in this case.  As a workaround, the user can
explicitly declare the output table as a `multiset` instead of
`relation`, as DDlog does not eliminate duplicates in multisets.

I just released DDlog v0.42.1, that adds a new switch to ovsdb2ddlog to
declare an output-only table as multiset.  We use this feature to
convert `Out_Logical_Flows` to a multiset.

Signed-off-by: Leonid Ryzhyk <lryzhyk@vmware.com>