ovn_northd.dl: UseLogicalDatapathGroups is back.
We bring back the `UseLogicalDatapathGroups` relation, used to enable
grouping of logical datapaths, but we do so in a way that only
introduces a single relatively small arrangements instead of two large
ones as was the case in an earlier implementation of this feature.
Specifically, we always unconditionally aggregate flows per datapath
and join the result with `UseLogicalDatapathGroups`, but only for flows
that belong to more than one datapath.  These flows must either be
flattened into one-flow-per-datapath if
`UseLogicalDatapathGroups[false]` or assigned to a datapath group when
`UseLogicalDatapathGroups[true]`.  This is a much smaller set of flows
that the entire `Flow` relation.  Besides, the new rules should only
require a single arrangement of these flows.

Signed-off-by: Leonid Ryzhyk <lryzhyk@vmware.com>